rename: FakeXXXProvider 전체 util 클래스로 설정
